Rename observability internal package name
