+++ .github/workflows/publish-release.yml
@@ -22,7 +22,7 @@ jobs:
         env:
           packageUser: ${{ github.actor }}
           packagePAT: ${{ secrets.GITHUB_TOKEN }}
-        run: ./gradlew build -x check -x test
+        run: ./gradlew build -x check -x test --stacktrace --console=plain
       - name: Run Trivy vulnerability scanner
         uses: aquasecurity/trivy-action@master
         with:
